Menu Administrativo
- Se Agrego proteccion de rutas al menu de admin
- Correccion de variable en detalle permiso
- Se validacion de vista por rol

+++ app/Http/Controllers/Admin/PermisosController.php
@@ -7,19 +7,23 @@ use Illuminate\Support\Facades\DB;
 use HelpDesk\Entities\Admin\Permission;
 use HelpDesk\Http\Controllers\Controller;
 use Symfony\Component\HttpFoundation\Response;
-use HelpDesk\Http\Requests\Admin\Permisos\CreatePermisoRequest;
-use HelpDesk\Http\Requests\Admin\Permisos\UpdatePermisoRequest;
+use HelpDesk\Http\Requests\Admin\Permisos\{CreatePermisoRequest, UpdatePermisoRequest};
+use Entrust;
 
 class PermisosController extends Controller
 {
     public function index()
     {
+        abort_unless(Entrust::can('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permisos = Permission::paginate();
         return view('admin.permisos.index', compact('permisos'));
     }
 
     public function create()
     {
+        abort_unless(Entrust::can('permission_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permiso = new Permission();
         return view('admin.permisos.create', compact('permiso'));
     }
@@ -32,22 +36,22 @@ class PermisosController extends Controller
             DB::commit();
 
             return redirect()->route('admin.permisos.index')
-            ->with([
-                'message' => 'Permiso Creado correctamente'
-            ]);
-
+                ->with([
+                    'message' => 'Permiso Creado correctamente'
+                ]);
         } catch (\Exception $e) {
             DB::rollback();
 
             return redirect()->back()
-            ->with([
-                'error' => 'Error Servidor; ' . $e->getMessage(),
-            ])->withInput();
+                ->with([
+                    'error' => 'Error Servidor; ' . $e->getMessage(),
+                ])->withInput();
         }
     }
 
     public function edit(Permission $permiso)
     {
+        abort_unless(Entrust::can('permission_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('admin.permisos.edit', compact('permiso'));
     }
 
@@ -62,25 +66,25 @@ class PermisosController extends Controller
             return  redirect()->route('admin.permisos.index')->with([
                 'message' => 'Permiso actualizado correctamente'
             ]);
-
-        }catch(\Exception $e){
+        } catch (\Exception $e) {
             DB::rollback();
 
             return redirect()->back()
-            ->with([
-                'error' => 'Error Servidor; ' . $e->getMessage(),
-            ])->withInput();
+                ->with([
+                    'error' => 'Error Servidor; ' . $e->getMessage(),
+                ])->withInput();
         }
-
     }
 
     public function show(Permission $permiso)
     {
+        abort_unless(Entrust::can('permission_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('admin.permisos.show', compact('permiso'));
     }
 
     public function destroy(Permission $permiso)
     {
+        abort_unless(Entrust::can('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $permiso->delete();
 
         return back();
